backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient